Initial implementation of Hack assembly instruction parsing
Implementation uses a deterministic finite automata (DFA) as the parsing algorithm. Data buffers provide additional memory for use by the DFA to store instruction data during the parsing. The DFA state is used to determine the outcome of the parsing; the exact nature of success (what instruction) or failure (what error). The for loop pushes character data through the DFA; implemented as the first match expression. The second match expression analyses the final state of the DFA to determine and assemble parsing outcome.