Added the ability to peek the next token
This comes at the cost of potential deep clones of StrConsts and Identifiers that overflow the internal (23 byte stack) buffer of CompactString. Should not be a major performance issue since these events will likely be relatively rare; identifiers will likely be shorter than 23 bytes, and most tokens will not be string constants or identifiers.